scanner: change OPCODE constants to u16
This also fixes the event_created_child macro to match a `pat` to allow paths and expressions to work.

+++ wayland-scanner/tests/scanner_assets/test-client-code.rs
@@ -42,19 +42,19 @@ pub mod wl_display {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_SYNC_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_SYNC_OPCODE: u32 = 0u32;
+    pub const REQ_SYNC_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_GET_REGISTRY_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_GET_REGISTRY_OPCODE: u32 = 1u32;
+    pub const REQ_GET_REGISTRY_OPCODE: u16 = 1u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_ERROR_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_ERROR_OPCODE: u32 = 0u32;
+    pub const EVT_ERROR_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_DELETE_ID_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_DELETE_ID_OPCODE: u32 = 1u32;
+    pub const EVT_DELETE_ID_OPCODE: u16 = 1u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -264,15 +264,15 @@ pub mod wl_registry {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_BIND_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_BIND_OPCODE: u32 = 0u32;
+    pub const REQ_BIND_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_GLOBAL_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_GLOBAL_OPCODE: u32 = 0u32;
+    pub const EVT_GLOBAL_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_GLOBAL_REMOVE_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_GLOBAL_REMOVE_OPCODE: u32 = 1u32;
+    pub const EVT_GLOBAL_REMOVE_OPCODE: u16 = 1u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -459,7 +459,7 @@ pub mod wl_callback {
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_DONE_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_DONE_OPCODE: u32 = 0u32;
+    pub const EVT_DONE_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {}
@@ -577,43 +577,43 @@ pub mod test_global {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_MANY_ARGS_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_MANY_ARGS_OPCODE: u32 = 0u32;
+    pub const REQ_MANY_ARGS_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_GET_SECONDARY_SINCE: u32 = 2u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_GET_SECONDARY_OPCODE: u32 = 1u32;
+    pub const REQ_GET_SECONDARY_OPCODE: u16 = 1u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_GET_TERTIARY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_GET_TERTIARY_OPCODE: u32 = 2u32;
+    pub const REQ_GET_TERTIARY_OPCODE: u16 = 2u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_LINK_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_LINK_OPCODE: u32 = 3u32;
+    pub const REQ_LINK_OPCODE: u16 = 3u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 4u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 4u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 4u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_REVERSE_LINK_SINCE: u32 = 5u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_REVERSE_LINK_OPCODE: u32 = 5u32;
+    pub const REQ_REVERSE_LINK_OPCODE: u16 = 5u16;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_NEWID_AND_ALLOW_NULL_SINCE: u32 = 5u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_NEWID_AND_ALLOW_NULL_OPCODE: u32 = 6u32;
+    pub const REQ_NEWID_AND_ALLOW_NULL_OPCODE: u16 = 6u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_MANY_ARGS_EVT_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_MANY_ARGS_EVT_OPCODE: u32 = 0u32;
+    pub const EVT_MANY_ARGS_EVT_OPCODE: u16 = 0u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_ACK_SECONDARY_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_ACK_SECONDARY_OPCODE: u32 = 1u32;
+    pub const EVT_ACK_SECONDARY_OPCODE: u16 = 1u16;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_CYCLE_QUAD_SINCE: u32 = 1u32;
     #[doc = r" The wire opcode for this event"]
-    pub const EVT_CYCLE_QUAD_OPCODE: u32 = 2u32;
+    pub const EVT_CYCLE_QUAD_OPCODE: u16 = 2u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -1072,7 +1072,7 @@ pub mod secondary {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 2u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -1196,7 +1196,7 @@ pub mod tertiary {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
@@ -1320,7 +1320,7 @@ pub mod quad {
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_DESTROY_SINCE: u32 = 3u32;
     #[doc = r" The wire opcode for this request"]
-    pub const REQ_DESTROY_OPCODE: u32 = 0u32;
+    pub const REQ_DESTROY_OPCODE: u16 = 0u16;
     #[derive(Debug)]
     #[non_exhaustive]
     pub enum Request {
